Remove OLS3 client, just use Ols4Client for everything
