+++ src/ols_py/__init__.py
@@ -1,4 +1,3 @@
-from .client import OlsClient
-from .ols4_client import Ols4Client
+from .client import Ols4Client
 
-__all__ = ["OlsClient", "Ols4Client"]
+__all__ = ["Ols4Client"]
